Move LongCollectionSqlType back to io.cassandrareaper.core
