+++ src/server/src/main/java/io/cassandrareaper/core/RepairSchedule.java
@@ -17,14 +17,14 @@
 
 package io.cassandrareaper.core;
 
-import io.cassandrareaper.storage.postgresql.LongCollectionSqlType;
-
+import java.util.Collection;
 import java.util.List;
 import java.util.UUID;
 import java.util.stream.Collectors;
 
 import com.google.common.base.Preconditions;
 import com.google.common.collect.ImmutableList;
+import com.google.common.collect.Lists;
 import org.apache.cassandra.repair.RepairParallelism;
 import org.joda.time.DateTime;
 
@@ -37,7 +37,8 @@ public final class RepairSchedule {
   private final int daysBetween;
   private final DateTime nextActivation;
   private final ImmutableList<UUID> runHistory;
-  @Deprecated private final int segmentCount;
+  @Deprecated
+  private final int segmentCount;
   private final RepairParallelism repairParallelism;
   private final double intensity;
   private final DateTime creationTime;
@@ -94,7 +95,8 @@ public final class RepairSchedule {
   }
 
   /**
-   * Required for JDBI mapping into database. Generic collection type would be hard to map into Postgres array types.
+   * Required for JDBI mapping into database. Generic collection type would be
+   * hard to map into Postgres array types.
    */
   public LongCollectionSqlType getRunHistorySql() {
     List<Long> list = runHistory.stream().map(UUID::getMostSignificantBits).collect(Collectors.toList());
@@ -139,9 +141,7 @@ public final class RepairSchedule {
   }
 
   public enum State {
-    ACTIVE,
-    PAUSED,
-    DELETED
+    ACTIVE, PAUSED, DELETED
   }
 
   public static final class Builder {
@@ -151,7 +151,8 @@ public final class RepairSchedule {
     private Integer daysBetween;
     private DateTime nextActivation;
     private ImmutableList<UUID> runHistory = ImmutableList.<UUID>of();
-    @Deprecated private int segmentCount = 0;
+    @Deprecated
+    private int segmentCount = 0;
     private RepairParallelism repairParallelism;
     private Double intensity;
     private DateTime creationTime = DateTime.now();
@@ -242,6 +243,23 @@ public final class RepairSchedule {
       Preconditions.checkState(null != intensity, "intensity(..) must be called before build(..)");
       Preconditions.checkState(null != segmentCountPerNode, "segmentCountPerNode(..) must be called before build(..)");
       return new RepairSchedule(this, id);
+    }
+  }
+
+  /**
+   * This is required to be able to map in generic manner into Postgres array
+   * types through JDBI.
+   */
+  public static final class LongCollectionSqlType {
+
+    private final Collection<Long> collection;
+
+    public LongCollectionSqlType(Collection<Long> collection) {
+      this.collection = collection;
+    }
+
+    public Collection<Long> getValue() {
+      return null != collection ? collection : Lists.newArrayList();
     }
   }
 }
