--- src/server/src/test/java/io/cassandrareaper/storage/LongCollectionSqlTypeTest.java
@@ -1,37 +1,0 @@
-/*
- *
- * Copyright 2019-2019 The Last Pickle Ltd
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package io.cassandrareaper.storage;
-
-import io.cassandrareaper.storage.postgresql.LongCollectionSqlType;
-
-import java.util.Arrays;
-import java.util.List;
-
-import org.junit.Test;
-
-import static org.junit.Assert.assertEquals;
-
-public final class LongCollectionSqlTypeTest {
-
-  @Test
-  public void testGetValue() {
-    List<Long> list = Arrays.asList(1L, 2L, 3L, 4L, 5L);
-    LongCollectionSqlType collection = new LongCollectionSqlType(list);
-    assertEquals(list, collection.getValue());
-  }
-}
